[refactor] Move boost txmempool definitions to their own header
Introduces mempool_set_definitions.h which collects the boost types that
were previously defined in the txmempool header.

The context of this change is an effort towards removing the boost multi
index includes from the header tree. The goal is that the experimental
`bitcoin-chainstate` binary can be compiled without requiring boost
includes.
