+++ src/node/miner.h
@@ -106,15 +106,15 @@ typedef boost::multi_index_container<
         // sorted by modified ancestor fee rate
         boost::multi_index::ordered_non_unique<
             // Reuse same tag from CTxMemPool's similar index
-            boost::multi_index::tag<ancestor_score>,
+            boost::multi_index::tag<MemPoolMultiIndex::ancestor_score>,
             boost::multi_index::identity<CTxMemPoolModifiedEntry>,
-            CompareTxMemPoolEntryByAncestorFee
+            MemPoolMultiIndex::CompareTxMemPoolEntryByAncestorFee
         >
     >
 > indexed_modified_transaction_set;
 
 typedef indexed_modified_transaction_set::nth_index<0>::type::iterator modtxiter;
-typedef indexed_modified_transaction_set::index<ancestor_score>::type::iterator modtxscoreiter;
+typedef indexed_modified_transaction_set::index<MemPoolMultiIndex::ancestor_score>::type::iterator modtxscoreiter;
 
 struct update_for_parent_inclusion
 {
