+++ src/txmempool.cpp
@@ -32,6 +32,9 @@
 #include <utility>
 
 using setEntries = std::set<CTxMemPool::txiter, CompareIteratorByHash>;
+using MemPoolMultiIndex::CompareTxMemPoolEntryByScore;
+using MemPoolMultiIndex::descendant_score;
+using MemPoolMultiIndex::entry_time;
 
 bool TestLockPointValidity(CChain& active_chain, const LockPoints& lp)
 {
